ManageAddress_BLService:: minor change

+++ src/main/java/com/sorted/portal/bl_services/ManageAddress_BLService.java
@@ -245,6 +245,9 @@ public class ManageAddress_BLService {
             if (address == null) {
                 throw new CustomIllegalArgumentsException(ResponseCode.NO_RECORD);
             }
+            if(address.isDeleted()){
+                return SEResponse.getEmptySuccessResponse(ResponseCode.ADD_ALREADY_DELETED);
+            }
             address.setDeleted(true);
             address_Service.deleteOne(address.getId(), usersBean.getId());
             return SEResponse.getEmptySuccessResponse(ResponseCode.SUCCESSFUL);
